% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gut_tree}
\alias{gut_tree}
\title{Phylogeny for the simulated gut MAGs}
\format{
A \code{phylo} object (see the \pkg{ape} package) with 24 tips whose
labels match the rows of \link{gut_counts}.
}
\usage{
gut_tree
}
\description{
An ultrametric coalescent tree over the 24 MAGs in \link{gut_counts}, scaled to
unit depth. Use it for the phylogenetic-diversity paths.
}
\examples{
hilldiv(gut_counts, q = c(0, 1), tree = gut_tree)
}
\seealso{
\link{gut_counts}, \link{gut_traits}
}
\keyword{datasets}
