% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hilldiv3-package.R
\docType{package}
\name{hilldiv3-package}
\alias{hilldiv3}
\alias{hilldiv3-package}
\title{hilldiv3: Integral Analysis of Diversity Based on Hill Numbers}
\description{
Tools for analysing the diversity of biological communities (e.g. OTU/ASV/MAG tables) based on Hill numbers. 'hilldiv3' provides a unified framework for neutral, phylogenetic and functional diversity measurement, diversity partitioning, (dis)similarity measurement, diversity profiles, evenness and redundancy. The statistical framework encompasses richness, Shannon and Simpson diversity, Faith's PD, Rao's quadratic entropy and Sorensen/Unifrac-type dissimilarities, all grounded in a single Hill-number framework. Methods are described in Jost (2007) \doi{10.1890/06-1736.1}, Chao et al. (2010) \doi{10.1098/rstb.2010.0272}, Chiu et al. (2014) \doi{10.1890/12-0960.1} and reviewed in Alberdi & Gilbert (2019) \doi{10.1111/1755-0998.13014}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/alberdilab/hilldiv3}
  \item \url{https://alberdilab.github.io/hilldiv3/}
  \item Report bugs at \url{https://github.com/alberdilab/hilldiv3/issues}
}

}
\author{
\strong{Maintainer}: Antton Alberdi \email{antton.alberdi@sund.ku.dk} (\href{https://orcid.org/0000-0002-2875-6446}{ORCID})

}
\keyword{internal}
